+++ registry/store/cluster/metadata.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/opencontainers/go-digest"
 	"github.com/robinkb/cascade-registry/cluster"
-	"github.com/robinkb/cascade-registry/store"
+	"github.com/robinkb/cascade-registry/registry/store"
 )
 
 func NewMetadataStore(proposer cluster.Proposer, metadata store.Metadata) store.Metadata {
