+++ registry/store/cluster/proposal_types.go
@@ -3,7 +3,7 @@ package cluster
 import (
 	"github.com/gofrs/uuid/v5"
 	"github.com/opencontainers/go-digest"
-	"github.com/robinkb/cascade-registry/store"
+	"github.com/robinkb/cascade-registry/registry/store"
 )
 
 // These types implement the cluster.Proposal interface.
