+++ registry/store/errors.go
