+++ registry/store/metadata.go
