+++ repository/manifests.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/opencontainers/go-digest"
 	v1 "github.com/opencontainers/image-spec/specs-go/v1"
-	"github.com/robinkb/cascade-registry/store"
+	"github.com/robinkb/cascade-registry/registry/store"
 )
 
 func (s *repositoryService) StatManifest(repository, id string) (*store.BlobInfo, error) {
