+++ repository/manifests_test.go
@@ -3,7 +3,7 @@ package repository
 import (
 	"testing"
 
-	"github.com/robinkb/cascade-registry/store"
+	"github.com/robinkb/cascade-registry/registry/store"
 	. "github.com/robinkb/cascade-registry/testing"
 )
