+++ repository/service.go
@@ -4,7 +4,7 @@ import (
 	"io"
 
 	"github.com/opencontainers/go-digest"
-	"github.com/robinkb/cascade-registry/store"
+	"github.com/robinkb/cascade-registry/registry/store"
 )
 
 type (
