+++ repository/service_test.go
@@ -3,10 +3,10 @@ package repository
 import (
 	"testing"
 
-	"github.com/robinkb/cascade-registry/store"
-	"github.com/robinkb/cascade-registry/store/boltdb"
-	"github.com/robinkb/cascade-registry/store/fs"
-	"github.com/robinkb/cascade-registry/store/inmemory"
+	"github.com/robinkb/cascade-registry/registry/store"
+	"github.com/robinkb/cascade-registry/registry/store/boltdb"
+	"github.com/robinkb/cascade-registry/registry/store/fs"
+	"github.com/robinkb/cascade-registry/registry/store/inmemory"
 	. "github.com/robinkb/cascade-registry/testing"
 	"github.com/stretchr/testify/suite"
 )
