+++ testing/conformance/push_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/robinkb/cascade-registry/registry"
 	v2 "github.com/robinkb/cascade-registry/registry/api/v2"
-	"github.com/robinkb/cascade-registry/store/inmemory"
+	"github.com/robinkb/cascade-registry/registry/store/inmemory"
 	. "github.com/robinkb/cascade-registry/testing"
 	testclient "github.com/robinkb/cascade-registry/testing/client"
 )
